role authorization

+++ fullstack/server/Api.Rest/Controllers/BookingController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 namespace Api.Rest.Controllers;
 
 [ApiController]
-public class BookingController(IBookingService bookingService) : ControllerBase
+public class BookingController(IBookingService bookingService, ISecurityService security) : ControllerBase
 {
     public const string ControllerRoute = "api/booking/";
     
@@ -16,15 +16,25 @@ public class BookingController(IBookingService bookingService) : ControllerBase
     
     [HttpPost]
     [Route(CreateBookingRoute)]
-    public ActionResult<BookingResponseDto> CreateBooking([FromBody] BookingCreateRequestDto dto)
+    public ActionResult<BookingResponseDto> CreateBooking([FromBody] BookingCreateRequestDto dto, [FromHeader] string authorization)
     {
+        var jwt = security.VerifyJwtOrThrow(authorization);
+        if (jwt.Role != "User")
+        {
+            return Unauthorized();
+        }
         return Ok(bookingService.CreateBooking(dto));
     }
     
     [HttpDelete]
     [Route(DeleteBookingRoute)]
-    public ActionResult<BookingResponseDto> DeleteBooking(string id)
+    public ActionResult<BookingResponseDto> DeleteBooking(string id, [FromHeader] string authorization)
     {
+        var jwt = security.VerifyJwtOrThrow(authorization);
+        if (jwt.Role!= "User")
+        {
+            return Unauthorized();
+        }
         return Ok(bookingService.DeleteBooking(id));
     }
 }
